wip: add users component
