+++ client/expenseus-fe/components/__tests__/Users.test.tsx
@@ -1,0 +1,10 @@
+import Users from "components/Users";
+import { render, screen } from "tests/test-utils";
+
+describe("Users component", () => {
+  it("should render a Users heading", () => {
+    render(<Users />);
+
+    expect(screen.getByText("Users")).toBeInTheDocument();
+  });
+});
