+++ client/expenseus-fe/pages/index.tsx
@@ -1,20 +1,6 @@
 import Head from "next/head";
 import "tailwindcss/tailwind.css";
 
-async function createUser() {
-  // make a post request to the URL
-}
-
-async function listUsers() {
-  // make a request to the URL
-  // return users
-}
-
-async function handleSubmit() {
-  await createUser();
-  await listUsers();
-}
-
 export default function Home() {
   return (
     <div className="container mx-auto px-4">
@@ -26,10 +12,6 @@ export default function Home() {
 
       <main className="py-4">
         <h1 className="text-4xl">Welcome to Expenseus</h1>
-        <section>
-          <h2>Users</h2>
-          <form onSubmit={handleSubmit}></form>
-        </section>
       </main>
     </div>
   );
